pyproject.toml
[project]
name = "threat-intel-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
